Added file ratios panel
This file ratios panel displays the character-to-word ratio of each document. Currently, there is a bug where, although most of the test files have a ratio of 6-8, there exists a file which has a character to word ratio of 65.

Bug testing reveals that it is because the Scanner is taking entire lines and interpreting it as a single word. This bug has not been fixed yet.